Move achievements viewer out of the achievements/viewer directory and into the top level.
