+++ test_game/main.lua
@@ -68,7 +68,7 @@ end
 import "CoreLibs/ui"
 import "generate_data"
 import "simple_viewer"
-import "achievements/viewer/achievements_viewer"
+import "achievements/viewer"
 
 local main_screen = playdate.ui.gridview.new(0, 20)
 local options = {
